+++ packages/sources/src/components/autosuggest/autosuggest.args.ts
@@ -6,6 +6,7 @@ export interface AutosuggestArgs {
   suggestOnFocus: boolean;
   loading: boolean;
   loadingLabel: string;
+  notFoundLabel: string;
   onSelect: HandlerFunction;
   onChange: HandlerFunction;
   onSearch: HandlerFunction;
@@ -23,6 +24,11 @@ export const autosuggestArgTypes: ArgTypes<AutosuggestArgs> = {
     },
   },
   loadingLabel: {
+    control: {
+      type: 'text',
+    },
+  },
+  notFoundLabel: {
     control: {
       type: 'text',
     },